refactor code + add citas by tipos
